Completed reorganization of Fractals folder  - Move all photos and code of original Markus-Lyapunov fractal into their own folder  - Update Markus-Lyapunov README to show the stochastic and probabilistic versions of the fractal (used to be in Fractals README)  - Update Fractals README to only show the basic form of each fractal -- now each folder holds the detailed pictures and code
