+++ Fractals/Markus-Lyapunov Fractals/README.md
@@ -1,0 +1,19 @@
+We have recreated the original Markus-Lyapunov fractal ([https://en.wikipedia.org/wiki/Lyapunov_fractal](https://en.wikipedia.org/wiki/Lyapunov_fractal)) in python with matplotlib to explore its properties.
+
+Stochastic AB Lyapunov fractal — in sequence "AB" there is a 50% chance at any time that the value will switch to the other (i.e., that A will switch to B and vice versa)
+![Stochastic Fractal Image](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_stochastic.png)
+
+Probabilistic AB Lyapunov fractal - in "AB" the probability that the value will switch varies is listed above the graph
+![Lyapunov_Fractal_Image_Prob_0](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_0.png)
+![Lyapunov_Fractal_Image_Prob_25](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_25.png)
+![Lyapunov_Fractal_Image_Prob_75](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_75.png)
+![Lyapunov_Fractal_Image_Prob_90](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_90.png)
+![Lyapunov_Fractal_Image_Prob_100](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_100.png)
+
+Animation (created using ImageMagick and ArtistAnimation) of the fractal changing with the probability
+![Animated Lyapunov Fractal Image](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/Prob_Lyap_Fractal_Avg.gif)
+
+
+
+
+
