+++ Fractals/Markus-Lyapunov Fractals/colored_lyapunov_fractal_ba.png
