+++ Fractals/Markus-Lyapunov Fractals/lyapunov_fractal_ab.png
