+++ Fractals/Markus-Lyapunov Fractals/lyapunov_fractal_probabilistic_AB_0.png
