+++ Fractals/Markus-Lyapunov Fractals/lyapunov_fractal_probabilistic_AB_100.png
