+++ Fractals/Markus-Lyapunov Fractals/lyapunov_fractal_stochastic.png
