+++ Fractals/README.md
@@ -1,17 +1,8 @@
-Here are pictures of various fractals, such as the Mandelbrot set or Julia set, and the Python code used to generate them.
+Here we explore various fractals and chaotic systems, such as the Markus-Lyapunov fractal or the double pendulum, using matplotlib and python. Each folder contains code and pictures/animations of our results for each different fractal. Below are the highlights of each folder.
 
-SuPeR Fractal from Poulsen, Ramsey, Schwartz paper
-![SuPeR Fractal Image](https://github.com/shaunramsey/FractalExploration/blob/master/Fractals/Lyapunov_Fractal_a_0.1.png)
+SuPeR fractal from Poulsen, Ramsey, & Schwartz paper
+![SuPeR Fractal Image](https://github.com/shaunramsey/FractalExploration/blob/master/Fractals/SuPeR Fractals/Lyapunov_Fractal_a_0.1.png)
 
-Stochastic AB Lyapunov fractal — in sequence "AB" there is a 50% chance at any time that the value will switch to the other (i.e., that A will switch to B and vice versa)
-![Stochastic Fractal Image](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_stochastic.png)
+Markus-Lyapunov fractal 
 
-Probabilistic AB Lyapunov fractal - in "AB" the probability that the value will switch varies is listed above the graph
-![Lyapunov_Fractal_Image_Prob_0](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_0.png)
-![Lyapunov_Fractal_Image_Prob_25](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_25.png)
-![Lyapunov_Fractal_Image_Prob_75](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_75.png)
-![Lyapunov_Fractal_Image_Prob_90](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_90.png)
-![Lyapunov_Fractal_Image_Prob_100](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/lyapunov_fractal_probabilistic_AB_100.png)
-
-Animation (created using ImageMagick and ArtistAnimation) of the fractal changing with the probability
-![Animated Lyapunov Fractal Image](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/Prob_Lyap_Fractal_Avg.gif)
+![Markus-Lyap Fractal Image](https://github.com/r-chambers/FractalExploration/blob/master/Fractals/Markus-Lyapunov Fractals/lyapunov_fractal_ab.png)
